Refactoring FileBaseGridCellRender
