+++ web-client/src/main/java/lsfusion/gwt/client/form/property/cell/classes/view/link/ImageLinkGridCellRenderer.java
@@ -1,11 +1,8 @@
 package lsfusion.gwt.client.form.property.cell.classes.view.link;
 
 import com.google.gwt.core.client.GWT;
-import com.google.gwt.dom.client.ImageElement;
 import lsfusion.gwt.client.form.property.GPropertyDraw;
 import lsfusion.gwt.client.form.property.cell.classes.view.ImageGridCellRenderer;
-
-import static lsfusion.gwt.client.form.property.cell.classes.view.FileGridCellRenderer.ICON_EMPTY;
 
 public class ImageLinkGridCellRenderer extends ImageGridCellRenderer {
     public ImageLinkGridCellRenderer(GPropertyDraw property) {
@@ -13,11 +10,7 @@ public class ImageLinkGridCellRenderer extends ImageGridCellRenderer {
     }
 
     @Override
-    protected void setFileSrc(ImageElement img, Object value) {
-        if (value instanceof String) {
-            img.setSrc((String) value);
-        } else {
-            img.setSrc(GWT.getModuleBaseURL() + ICON_EMPTY);
-        }
+    protected String prepareFilePath(Object value) {
+        return value instanceof String ? value.toString() : GWT.getModuleBaseURL() + ICON_EMPTY;
     }
 }
