+++ web-client/src/main/java/lsfusion/gwt/client/form/property/cell/view/FileBasedGridCellRenderer.java
@@ -1,5 +1,6 @@
 package lsfusion.gwt.client.form.property.cell.view;
 
+import com.google.gwt.dom.client.Document;
 import com.google.gwt.dom.client.Element;
 import com.google.gwt.dom.client.ImageElement;
 import com.google.gwt.dom.client.Style;
@@ -7,6 +8,9 @@ import lsfusion.gwt.client.form.design.GFont;
 import lsfusion.gwt.client.form.property.GPropertyDraw;
 
 public abstract class FileBasedGridCellRenderer extends AbstractGridCellRenderer {
+    protected static final String ICON_EMPTY = "empty.png";
+    protected static final String ICON_FILE = "file.png";
+
     protected GPropertyDraw property;
 
     @Override
@@ -19,6 +23,27 @@ public abstract class FileBasedGridCellRenderer extends AbstractGridCellRenderer
         if (!isSingle) element.getStyle().setPosition(Style.Position.RELATIVE);
     }
 
+    @Override
+    public void renderDynamic(Element element, GFont font, Object value, boolean isSingle) {
+        if (value == null && property.isEditableNotNull()) {
+            setBasedEmptyElement(element);
+        } else {
+            element.getStyle().clearPadding();
+            element.removeClassName("requiredValueString");
+            element.setInnerText(null);
+            element.setTitle("");
+
+            ImageElement img = element.appendChild(Document.get().createImageElement());
+
+            Style imgStyle = img.getStyle();
+            imgStyle.setVerticalAlign(Style.VerticalAlign.MIDDLE);
+            imgStyle.setProperty("maxWidth", "100%");
+            imgStyle.setProperty("maxHeight", "100%");
+
+            img.setSrc(prepareFilePath(value));
+        }
+    }
+
     protected void setBasedEmptyElement(Element element){
         element.getStyle().setPaddingRight(4, Style.Unit.PX);
         element.getStyle().setPaddingLeft(4, Style.Unit.PX);
@@ -27,9 +52,9 @@ public abstract class FileBasedGridCellRenderer extends AbstractGridCellRenderer
         element.addClassName("requiredValueString");
     }
 
+    protected abstract String prepareFilePath(Object value);
+
     protected FileBasedGridCellRenderer(GPropertyDraw property) {
         this.property = property;
     }
-
-    protected abstract void setFileSrc(ImageElement file, Object value);
 }
